netbox-interact/requirements.txt
pynetbox
pyyaml
ruamel.yaml
requests
